Added files from Skeleton 2 distribution version 02.01.03 (released 19 October 2010). The package was downloaded from http://www.uic.unn.ru/~zny/skeleton.
The files were added with moderate amount of renames and reorganization inside original directory organization. The content of files were added "as is" without reflection of those changes in the directory structure. So, building infrastructure has been broken for both Windows and Linux and needs to be restored.

Manual wasn't added due to absence of sources for it in the original package. Needs to be added.

Added files likely violate ACG. Needs to be fixed.

Think about adding Matlab wrapper.